move modern Layout
