+++ src/components/HistoryEventDetails/layouts/TokenExchangeEventDetails.js
@@ -27,7 +27,7 @@ import { useDispatch } from 'react-redux';
 import { viewTransactionOnBlockchainAction } from 'actions/historyActions';
 
 // Components
-import { Row, ColumnRight, Spacing } from 'components/modern/Layout';
+import { Row, ColumnRight, Spacing } from 'components/layout/Layout';
 import Button from 'components/core/Button';
 import FeeLabel from 'components/modern/FeeLabel';
 import TokenValueView from 'components/modern/TokenValueView';
