+++ src/components/HistoryEventDetails/layouts/TokenTransactionEventDetails.js
@@ -27,7 +27,7 @@ import { useDispatch } from 'react-redux';
 import { viewTransactionOnBlockchainAction } from 'actions/historyActions';
 
 // Components
-import { Row, Spacing } from 'components/modern/Layout';
+import { Row, Spacing } from 'components/layout/Layout';
 import Button from 'components/core/Button';
 import FeeLabel from 'components/modern/FeeLabel';
 import TokenValueView from 'components/modern/TokenValueView';
