+++ src/components/HistoryList/items/TokenExchangeItem.js
@@ -23,7 +23,7 @@ import { useTranslation } from 'translations/translate';
 
 // Components
 import TokenValueView from 'components/modern/TokenValueView';
-import { ColumnRight } from 'components/modern/Layout';
+import { ColumnRight } from 'components/layout/Layout';
 
 // Types
 import type { TokenExchangeEvent } from 'models/History';
