+++ src/components/Modals/PrismicDocumentModal/PrismicDocumentModal.js
@@ -26,7 +26,7 @@ import t from 'translations/translate';
 // Components
 import Modal from 'components/Modal';
 import HeaderBlock from 'components/HeaderBlock';
-import { Container } from 'components/modern/Layout';
+import { Container } from 'components/layout/Layout';
 import Text from 'components/core/Text';
 
 // Types
