+++ src/components/inputs/TokenFiatValueInputs.js
@@ -25,7 +25,7 @@ import styled from 'styled-components/native';
 import t from 'translations/translate';
 
 // Components
-import { Spacing } from 'components/modern/Layout';
+import { Spacing } from 'components/layout/Layout';
 import TokenValueInput from 'components/inputs/TokenValueInput';
 import TokenBalanceAccessory from 'components/inputs/TokenValueInput/TokenBalanceAccessory';
 import FiatValueInput from 'components/inputs/FiatValueInput';
