+++ src/screens/AddCash/AddCash.js
@@ -36,7 +36,7 @@ import { useThemeColors } from 'utils/themes';
 import { isLogV2AppEvents } from 'utils/environment';
 
 // Components
-import { Container } from 'components/modern/Layout';
+import { Container } from 'components/layout/Layout';
 import Button from 'components/core/Button';
 import TextInput from 'components/legacy/TextInput';
 import HeaderBlock from 'components/HeaderBlock';
