+++ src/screens/Asset/Asset.js
@@ -25,7 +25,7 @@ import t from 'translations/translate';
 import { useNavigation, useNavigationParam } from 'react-navigation-hooks';
 
 // Components
-import { Container, Content } from 'components/modern/Layout';
+import { Container, Content } from 'components/layout/Layout';
 import HeaderBlock from 'components/HeaderBlock';
 import FloatingButtons from 'components/FloatingButtons';
 import ActivityFeed from 'components/legacy/ActivityFeed';
