+++ src/screens/Exchange/ExchangeConfirm/ExchangeConfirm.js
@@ -24,7 +24,7 @@ import styled from 'styled-components/native';
 import { useTranslation } from 'translations/translate';
 
 // Components
-import { Container, Content, Spacing } from 'components/modern/Layout';
+import { Container, Content, Spacing } from 'components/layout/Layout';
 import HeaderBlock from 'components/HeaderBlock';
 import Button from 'components/core/Button';
 import Text from 'components/core/Text';
