+++ src/screens/KeyBasedAssetTransfer/KeyBasedAssetTransferIntro/KeyBasedAssetTransferIntro.js
@@ -32,7 +32,7 @@ import { KEY_BASED_ASSET_TRANSFER_CHOOSE } from 'constants/navigationConstants';
 // Components
 import { Spacing } from 'components/legacy/Layout';
 import Button from 'components/legacy/Button';
-import { Container, Content } from 'components/modern/Layout';
+import { Container, Content } from 'components/layout/Layout';
 import HeaderBlock from 'components/HeaderBlock';
 import Text from 'components/core/Text';
 import EmojiCard from 'components/modern/EmojiCard';
