+++ src/screens/Menu/SelectLanguage.js
@@ -29,7 +29,7 @@ import { useTranslationWithPrefix } from 'translations/translate';
 import localeConfig from 'configs/localeConfig';
 
 // Components
-import { Container } from 'components/modern/Layout';
+import { Container } from 'components/layout/Layout';
 import HeaderBlock from 'components/HeaderBlock';
 import Icon from 'components/core/Icon';
 import Text from 'components/core/Text';
