+++ src/screens/WalletMigrationArchanova/Review/WalletMigrationArchanovaReview.js
@@ -23,7 +23,7 @@ import { useNavigation } from 'react-navigation-hooks';
 import { useTranslationWithPrefix } from 'translations/translate';
 
 // Components
-import { Container, Content, Footer } from 'components/modern/Layout';
+import { Container, Content, Footer } from 'components/layout/Layout';
 import Button from 'components/core/Button';
 import CheckAuth from 'components/CheckAuth';
 import HeaderBlock from 'components/HeaderBlock';
