+++ src/screens/WalletMigrationArchanova/SelectAssets/WalletMigrationArchanovaSelectAssets.js
@@ -26,7 +26,7 @@ import { BigNumber } from 'bignumber.js';
 import { isEmpty } from 'lodash';
 
 // Components
-import { Container, Footer } from 'components/modern/Layout';
+import { Container, Footer } from 'components/layout/Layout';
 import Button from 'components/core/Button';
 import HeaderBlock from 'components/HeaderBlock';
